requirements.txt
python-dotenv>=1.0.0
openai>=1.2.4
pyaudio
elevenlabs